feat: linked list in C

+++ dsa/binary-search/c/main.c
@@ -2,9 +2,9 @@
 
 #define CMP(a, b) (((a) < (b)) ? -1 : ((a) == (b)) ? 0 : 1)
 
-int bin_search(int arr[], int left, int right, int target) {
+int bin_search(int arr[], size_t left, size_t right, int target) {
   while (left <= right) {
-    int mid = (left + right) / 2;
+    size_t mid = (left + right) / 2;
     switch (CMP(target, arr[mid])) {
     case -1:
       right = mid - 1;
@@ -18,9 +18,9 @@ int bin_search(int arr[], int left, int right, int target) {
   return -1;
 }
 
-int bin_search_rec(int arr[], int left, int right, int target) {
+int bin_search_rec(int arr[], size_t left, size_t right, int target) {
   if (left <= right) {
-    int mid = (left + right) / 2;
+    size_t mid = (left + right) / 2;
     switch (CMP(target, arr[mid])) {
     case -1:
       return bin_search_rec(arr, left, mid - 1, target);
@@ -35,7 +35,7 @@ int bin_search_rec(int arr[], int left, int right, int target) {
 
 int main(void) {
   int arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-  int len = sizeof(arr) / sizeof(arr[0]);
+  size_t len = sizeof(arr) / sizeof(arr[0]);
   int target = 100;
   int result = bin_search_rec(arr, 0, len - 1, target);
   printf("Index of %d is %d\n", target, result);
